pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud-red-team-arena"
version = "1.0.0"
description = "A cloud security reinforcement learning environment where an AI agent performs red-teaming tasks including discovering vulnerabilities, exploiting them, and patching them."
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "requests",
    "openai",
    "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"
